changed double to float

+++ linalg/LinearAlgebra/mul.c
@@ -11,7 +11,7 @@
  * C with row_a x column_b
  */
 
-void mul(double* A, double* B, bool elementWise, double* C, int row_a, int column_a, int column_b) {
+void mul(float* A, float* B, bool elementWise, float* C, int row_a, int column_a, int column_b) {
 
     /*
      * C = A*B if elementWise = false
@@ -19,9 +19,9 @@ void mul(double* A, double* B, bool elementWise, double* C, int row_a, int colum
      */
 
     // Data matrix
-    double* data_a = A;
-    double* data_b = B;
-    // double* C = C; // No need
+    float* data_a = A;
+    float* data_b = B;
+    // float* C = C; // No need
 
     if (elementWise == true) {
 
@@ -76,7 +76,7 @@ void mulf(float* A, float* B, bool elementWise, float* C, int row_a, int column_
     // Data matrix
     float* data_a = A;
     float* data_b = B;
-    // double* C = C; // No need
+    // float* C = C; // No need
 
     if (elementWise == true) {
